+++ bigml/multivote.py
@@ -47,6 +47,43 @@ WEIGHT_KEYS = {
 DEFAULT_METHOD = 0
 
 
+def ws_confidence(prediction, distribution, ws_z=1.96, ws_n=None):
+    """Wilson score interval computation of the distribution for the prediction
+
+       expected arguments:
+            prediction: the value of the prediction for which confidence is
+                        computed
+            distribution: a distribution-like structure of predictions and
+                          the associated weights. (e.g.
+                          [['Iris-setosa', 10], ['Iris-versicolor', 5]])
+            ws_z: percentile of the standard normal distribution
+            ws_n: total number of instances in the distribution. If absent,
+                  the number is computed as the sum of weights in the
+                  provided distribution
+
+    """
+    if isinstance(distribution, list):
+        distribution = dict(distribution)
+    ws_p = distribution[prediction]
+    if ws_p < 0:
+        raise ValueError("The distribution weight must be a positive value")
+    if ws_n < 1:
+        raise ValueError("The total of instances in the distribution must be"
+                         " a positive integer")
+    ws_norm = float(sum(distribution.values()))
+    if not ws_norm == 1.0:
+        ws_p = ws_p / ws_norm
+    if ws_n is None:
+        ws_n = ws_norm
+    else:
+        ws_n = float(ws_n)
+        ws_z = float(ws_z)
+    ws_z2 = ws_z * ws_z
+    ws_factor = ws_z2 / ws_n
+    ws_sqrt = math.sqrt((ws_p * (1 - ws_p) + ws_factor / 4) / ws_n)
+    return (ws_p + ws_factor / 2 - ws_z * ws_sqrt) / (1 + ws_factor)
+
+
 class MultiVote(object):
     """A multiple vote prediction
 
@@ -55,24 +92,42 @@ class MultiVote(object):
     """
 
     @classmethod
-    def avg(cls, instance):
+    def avg(cls, instance, with_confidence=False):
         """Returns the average of a list of numeric values.
 
+           If with_confidence is True, the combined confidence (as the
+           average of confidences of the multivote predictions) is also
+           returned
         """
+        if (instance.predictions and with_confidence and
+                not all(['confidence' in prediction
+                         for prediction in instance.predictions])):
+            raise Exception("Not enough data to use the selected "
+                            "prediction method. Try creating your"
+                            " model anew.")
         total = len(instance.predictions)
         result = 0.0
+        confidence = 0.0
         for prediction in instance.predictions:
             result += prediction['prediction']
-        return result / total if total > 0 else float('nan')
+            if with_confidence:
+                confidence += prediction['confidence']
+        if with_confidence:
+            return ((result / total, confidence / total) if total > 0 else
+                    (float('nan'), 0))
+        return result if total > 0 else float('nan')
 
     @classmethod
-    def error_weighted(cls, instance):
+    def error_weighted(cls, instance, with_confidence=False):
         """Returns the prediction combining votes using error to compute weight
 
+           If with_confidences is true, the combined confidence (as the
+           error weighted average of the confidences of the multivote
+           predictions) is also returned
         """
-        if instance.predictions and not all(['confidence' in prediction
-                                             for prediction
-                                             in instance.predictions]):
+        if (instance.predictions and with_confidence and
+                not all(['confidence' in prediction
+                         for prediction in instance.predictions])):
             raise Exception("Not enough data to use the selected "
                             "prediction method. Try creating your"
                             " model anew.")
@@ -80,10 +135,22 @@ class MultiVote(object):
         result = 0.0
         normalization_factor = cls.normalize_error(instance, top_range)
         if normalization_factor == 0:
-            return float('nan')
+            if with_confidence:
+                return float('nan'), 0
+            else:
+                return float('nan')
+        if with_confidence:
+            combined_error = 0.0
         for prediction in instance.predictions:
             result += prediction['prediction'] * prediction['error_weight']
-        return result / normalization_factor
+            if with_confidence:
+                combined_error += (prediction['confidence'] *
+                                   prediction['error_weight'])
+        if with_confidence:
+            return (result / normalization_factor,
+                    combined_error / normalization_factor)
+        else:
+            return result / normalization_factor
 
     @classmethod
     def normalize_error(cls, instance, top_range):
@@ -120,12 +187,19 @@ class MultiVote(object):
     def __init__(self, predictions):
         """Init method, builds a MultiVote with a list of predictions
 
+           The constuctor expects a list of well formed predictions like:
+                {'prediction': 'Iris-setosa', 'confidence': 0.7}
+            Each prediction can also contain an 'order' key that is used
+            to break even in votations. The list order is used by default.
         """
         self.predictions = []
         if isinstance(predictions, list):
             self.extend(predictions)
         else:
             self.append(predictions)
+        if not all(['order' in prediction for prediction in predictions]):
+            for i in range(len(self.predictions)):
+                self.predictions[i]['order'] = i
 
     def is_regression(self):
         """Returns True if all the predictions are numbers
@@ -146,10 +220,15 @@ class MultiVote(object):
             return self.predictions[-1]['order'] + 1
         return 0
 
-    def combine(self, method=DEFAULT_METHOD):
+    def combine(self, method=DEFAULT_METHOD, with_confidence=False):
         """Reduces a number of predictions voting for classification and
            averaging predictions for regression.
 
+           method will determine the voting method (plurality, confidence
+           weighted or probability weighted).
+           If with_confidence is true, the combined confidence (as a weighted
+           average of the confidences of votes for the combined prediction)
+           will also be given.
         """
         # there must be at least one prediction to be combined
         if not self.predictions:
@@ -166,8 +245,9 @@ class MultiVote(object):
                                     "prediction method. Try creating your"
                                     " model anew.")
         if self.is_regression():
-            return NUMERICAL_COMBINATION_METHODS.get(method,
-                                                     self.__class__.avg)(self)
+            function = NUMERICAL_COMBINATION_METHODS.get(method,
+                                                         self.__class__.avg)
+            return function(self, with_confidence=with_confidence)
         else:
             if method == PROBABILITY:
                 predictions = MultiVote([])
@@ -175,7 +255,8 @@ class MultiVote(object):
             else:
                 predictions = self
             return predictions.combine_categorical(
-                COMBINATION_WEIGHTS.get(method, None))
+                COMBINATION_WEIGHTS.get(method, None),
+                with_confidence=with_confidence)
 
     def probability_weight(self):
         """Reorganizes predictions depending on training data probability
@@ -195,16 +276,47 @@ class MultiVote(object):
             for prediction, instances in prediction['distribution']:
                 predictions.append({'prediction': prediction,
                                     'probability': float(instances) / total,
+                                    'count': instances,
                                     'order': order})
         return predictions
 
-    def combine_categorical(self, weight_label=None):
+    def combine_distribution(self, weight_label='probability'):
+        """Builds a distribution based on the predictions of the MultiVote
+
+           Given the array of predictions, we build a set of predictions with
+           them and associate the sum of weights (the weight being the
+           contents of the weight_label field of each prediction)
+        """
+        if not all([weight_label in prediction
+                    for prediction in self.predictions]):
+            raise Exception("Not enough data to use the selected "
+                            "prediction method. Try creating your"
+                            " model anew.")
+        distribution = {}
+        total = 0
+        for prediction in self.predictions:
+            if not prediction['prediction'] in distribution:
+                distribution[prediction['prediction']] = 0.0
+            distribution[prediction['prediction']] += prediction[weight_label]
+            total += prediction['count']
+        if total > 0:
+            distribution = [[key, value] for key, value in
+                            distribution.items()]
+        else:
+            distribution = []
+        return distribution, total
+
+    def combine_categorical(self, weight_label=None, with_confidence=False):
         """Returns the prediction combining votes by using the given weight:
 
             weight_label can be set as:
             None:          plurality (1 vote per prediction)
             'confidence':  confidence weighted (confidence as a vote value)
             'probability': probability weighted (probability as a vote value)
+
+            If with_confidence is true, the combined confidence (as a weighted
+            average of the confidences of the votes for the combined
+            prediction) will also be given.
         """
         mode = {}
         if weight_label is None:
@@ -227,9 +339,48 @@ class MultiVote(object):
             else:
                 mode[category] = {"count": weight,
                                   "order": prediction['order']}
-        return sorted(mode.items(), key=lambda x: (x[1]['count'],
-                                                   -x[1]['order']),
-                      reverse=True)[0][0]
+
+        prediction = sorted(mode.items(), key=lambda x: (x[1]['count'],
+                                                         -x[1]['order']),
+                            reverse=True)[0][0]
+
+        if with_confidence:
+            if 'confidence' in self.predictions[0]:
+                return self.weighted_confidence(prediction,
+                                                weight_label)
+            # if prediction had no confidence, compute it from distribution
+            else:
+                combined_distribution = self.combine_distribution()
+                distribution, count = combined_distribution
+                combined_confidence = ws_confidence(prediction, distribution,
+                                                    ws_n=count)
+                return prediction, combined_confidence
+        return prediction
+
+    def weighted_confidence(self, combined_prediction, weight_label):
+        """Compute the combined weighted confidence from a list of predictions
+
+        """
+        predictions = filter(lambda x: x['prediction'] == combined_prediction,
+                             self.predictions)
+        if (weight_label is not None and
+                (not isinstance(weight_label, basestring) or
+                 any([not 'confidence' or not weight_label in prediction
+                 for prediction in predictions]))):
+            raise ValueError("Not enough data to use the selected "
+                             "prediction method. Lacks %s information." %
+                             weight_label)
+        final_confidence = 0.0
+        total_weight = 0.0
+        weight = 1
+        for prediction in predictions:
+            if weight_label is not None:
+                weight = prediction[weight_label]
+            final_confidence += weight * prediction['confidence']
+            total_weight += weight
+        final_confidence = (final_confidence / total_weight
+                            if total_weight > 0 else float('nan'))
+        return combined_prediction, final_confidence
 
     def append(self, prediction_info):
         """Adds a new prediction into a list of predictions